aiko_diffusion/dataset/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataset-aiko"
version = "0.1.0"
description = "Pixel art dataset augmentation & filtering toolkit for diffusion model fine-tuning"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "numpy>=1.24",
    "PyYAML>=6.0",
    "click>=8.1",
    "scikit-learn>=1.3",
    "torch>=2.0",
    "torchvision>=0.15",
    "imagehash>=4.3",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
dataset-aiko = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
